+++ ui/src/components/search-box/QasSearchBox.vue
@@ -74,7 +74,6 @@ export default {
     },
 
     outlined: {
-      default: false,
       type: Boolean
     },